fix: Use models passed with `-m` as-is
Previously, the `-m` switch acted as a filter, deciding which of the
git-tracked models to process. This led to very unintuitive behavior,
and placed some unnecessary restrictions on what could be passed this
way (only paths to `*.aird` files relative to the repo root).

Instead, now `-m` accepts an arbitrary path, and will try to load a
model from there without any further restrictions. This also means that
a model passed this way does not necessarily have to be inside a git
repository, if either `--no-commit` is passed or `--fix` is not passed.

The behavior when `-m` is not passed at all remains unchanged; in this
case the script will still run over all models tracked in the local git
repository.